Certbot and Tile Viewer
Merging remote work from local to avoid git confusion. Adding tile viewer which has been done in perhapse the worst way possible: a second version of the bingo tile invisibly rests on top of the real one, which contains the tile viewer. But it works, is super fast, and has a cute zoom function. Todo: add links to the card tiles that open the viewer to the description, add a 'deny' function to card tiles.